Программирует открытие/закрытие компонентов стрелками
