+++ src/view/event.js
@@ -49,9 +49,6 @@ function createEvent(point, offers, destination) {
                 <path d="M14 21l-8.22899 4.3262 1.57159-9.1631L.685209 9.67376 9.8855 8.33688 14 0l4.1145 8.33688 9.2003 1.33688-6.6574 6.48934 1.5716 9.1631L14 21z"/>
               </svg>
             </button>
-            <button class="event__rollup-btn" type="button">
-              <span class="visually-hidden">Open event</span>
-            </button>
           </div>`;
 }
 
@@ -59,6 +56,7 @@ export default class Event extends AbstractView {
   #point;
   #offers;
   #destination;
+
   constructor ({point, offers, destination}) {
     super();
     this.#point = point;
