+++ src/view/new-point.js
@@ -1,137 +1,164 @@
 import AbstractView from '../framework/view/abstract-view.js';
-import { humanizeDate, ucFirst, DATE_FORMAT_FIRST } from '../utils.js';
-import { TYPES } from '../mock/const.js';
 
-const BLANK_POINT = {
-  id: 1,
-  type: 'Flight',
-  offers: [],
-  destination: '',
-  dateFrom: '',
-  dateTo: '',
-  price: 0,
-};
-
-function createTypesList(types, type) {
-  return types.map((element) => `
-          <div class="event__type-item">
-            <input id="event-type-${element}-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="${element}"${element === type ? ' checked' : ''}>
-            <label class="event__type-label  event__type-label--${element}" for="event-type-${element}-1">${ucFirst(element)}</label>
-          </div>`).join('');
-}
-
-function createOffersTemplate(offers, pointsOffers) {
-  return offers.length > 0 ?
-    `<section class="event__section  event__section--offers">
-        <h3 class="event__section-title  event__section-title--offers">Offers</h3>
-        <div class="event__available-offers">
-        ${offers.map((offer) =>
-    `<div class="event__offer-selector">
-      <input class="event__offer-checkbox  visually-hidden" id="event-offer-${offer.id}-1" type="checkbox" name="event-offer-${offer.id}"${pointsOffers.includes(offer.id) ? ' checked' : ''}>
-      <label class="event__offer-label" for="event-offer-${offer.id}-1">
-        <span class="event__offer-title">${offer.title}</span>
-        &plus;&euro;&nbsp;
-        <span class="event__offer-price">${offer.price}</span>
-      </label>
-    </div>`).join('')}
-      </div>
-    </section>` : '';
-}
-
-function createDestinationTemplate({ name, description, photos }) {
-  const photosTemplate = createPhotosTemplate(photos);
-
-  return `<section class="event__section  event__section--destination">
-    <h3 class="event__section-title  event__section-title--destination">Destination</h3>
-    <p class="event__destination-description">${name}. ${description}</p>
-    ${photosTemplate}
-  </section>`;
-}
-
-function createPhotosTemplate(photos) {
-  return photos.length > 0 ?
-    `<div class="event__photos-container">
-      <div class="event__photos-tape">
-        ${photos.map((photo) => `<img class="event__photo" src="${photo.src}" alt="${photo.description}">`).join('')}
-      </div>
-    </div>` : '';
-}
-
-function createNewPoint(point, offers, destination) {
-  const { id, type, dateFrom, dateTo, price } = point;
-  const { name, description, photos } = typeof destination !== 'undefined' ? destination : '';
-  const destinationName = typeof name !== 'undefined' ? name : '';
-  const dateFromHumanized = humanizeDate(dateFrom, DATE_FORMAT_FIRST);
-  const dateToHumanized = humanizeDate(dateTo, DATE_FORMAT_FIRST);
-  const typesList = createTypesList(TYPES, type);
-  const offersTemplate = createOffersTemplate(offers, point.offers);
-  const destinationTemplate = typeof destination !== 'undefined' ? (createDestinationTemplate({ name, description, photos })) : '';
-
+function createNewPoint() {
   return `<form class="event event--edit" action="#" method="post">
             <header class="event__header">
               <div class="event__type-wrapper">
-                <label class="event__type  event__type-btn" for="event-type-toggle-${id}">
+                <label class="event__type  event__type-btn" for="event-type-toggle-1">
                   <span class="visually-hidden">Choose event type</span>
-                  <img class="event__type-icon" width="17" height="17" src="img/icons/${type}.png" alt="${type}">
+                  <img class="event__type-icon" width="17" height="17" src="img/icons/flight.png" alt="Event type icon">
                 </label>
-                <input class="event__type-toggle  visually-hidden" id="event-type-toggle-${id}" type="checkbox">
+                <input class="event__type-toggle  visually-hidden" id="event-type-toggle-1" type="checkbox">
 
                 <div class="event__type-list">
                   <fieldset class="event__type-group">
                     <legend class="visually-hidden">Event type</legend>
-                    ${typesList}
+
+                    <div class="event__type-item">
+                      <input id="event-type-taxi-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="taxi">
+                      <label class="event__type-label  event__type-label--taxi" for="event-type-taxi-1">Taxi</label>
+                    </div>
+
+                    <div class="event__type-item">
+                      <input id="event-type-bus-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="bus">
+                      <label class="event__type-label  event__type-label--bus" for="event-type-bus-1">Bus</label>
+                    </div>
+
+                    <div class="event__type-item">
+                      <input id="event-type-train-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="train">
+                      <label class="event__type-label  event__type-label--train" for="event-type-train-1">Train</label>
+                    </div>
+
+                    <div class="event__type-item">
+                      <input id="event-type-ship-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="ship">
+                      <label class="event__type-label  event__type-label--ship" for="event-type-ship-1">Ship</label>
+                    </div>
+
+                    <div class="event__type-item">
+                      <input id="event-type-drive-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="drive">
+                      <label class="event__type-label  event__type-label--drive" for="event-type-drive-1">Drive</label>
+                    </div>
+
+                    <div class="event__type-item">
+                      <input id="event-type-flight-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="flight" checked>
+                      <label class="event__type-label  event__type-label--flight" for="event-type-flight-1">Flight</label>
+                    </div>
+
+                    <div class="event__type-item">
+                      <input id="event-type-check-in-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="check-in">
+                      <label class="event__type-label  event__type-label--check-in" for="event-type-check-in-1">Check-in</label>
+                    </div>
+
+                    <div class="event__type-item">
+                      <input id="event-type-sightseeing-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="sightseeing">
+                      <label class="event__type-label  event__type-label--sightseeing" for="event-type-sightseeing-1">Sightseeing</label>
+                    </div>
+
+                    <div class="event__type-item">
+                      <input id="event-type-restaurant-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="restaurant">
+                      <label class="event__type-label  event__type-label--restaurant" for="event-type-restaurant-1">Restaurant</label>
+                    </div>
                   </fieldset>
                 </div>
               </div>
 
               <div class="event__field-group  event__field-group--destination">
                 <label class="event__label  event__type-output" for="event-destination-1">
-                  ${type}
+                  Flight
                 </label>
-                <input class="event__input  event__input--destination" id="event-destination-1" type="text" name="event-destination" value="${destinationName}" list="destination-list-1">
+                <input class="event__input  event__input--destination" id="event-destination-1" type="text" name="event-destination" value="Chamonix" list="destination-list-1">
                 <datalist id="destination-list-1">
-                  <option value="${destinationName}"></option>
+                  <option value="Amsterdam"></option>
+                  <option value="Geneva"></option>
+                  <option value="Chamonix"></option>
                 </datalist>
               </div>
 
               <div class="event__field-group  event__field-group--time">
                 <label class="visually-hidden" for="event-start-time-1">From</label>
-                <input class="event__input  event__input--time" id="event-start-time-1" type="text" name="event-start-time" value="${dateFromHumanized}">
+                <input class="event__input  event__input--time" id="event-start-time-1" type="text" name="event-start-time" value="18/03/19 12:25">
                 &mdash;
                 <label class="visually-hidden" for="event-end-time-1">To</label>
-                <input class="event__input  event__input--time" id="event-end-time-1" type="text" name="event-end-time" value="${dateToHumanized}">
+                <input class="event__input  event__input--time" id="event-end-time-1" type="text" name="event-end-time" value="18/03/19 13:35">
               </div>
 
               <div class="event__field-group  event__field-group--price">
                 <label class="event__label" for="event-price-1">
                   <span class="visually-hidden">Price</span>
-                  &euro; ${price}
+                  &euro;
                 </label>
-                <input class="event__input  event__input--price" id="event-price-1" type="text" name="event-price" value="">
+                <input class="event__input  event__input--price" id="event-price-1" type="text" name="event-price" value="160">
               </div>
 
               <button class="event__save-btn  btn  btn--blue" type="submit">Save</button>
-              <button class="event__reset-btn" type="reset">Cancel</button>
+              <button class="event__reset-btn" type="reset">Delete</button>
+              <button class="event__rollup-btn" type="button">
+                <span class="visually-hidden">Open event</span>
+              </button>
             </header>
             <section class="event__details">
-              ${offersTemplate}
-              ${destinationTemplate}
+              <section class="event__section  event__section--offers">
+                <h3 class="event__section-title  event__section-title--offers">Offers</h3>
+
+                <div class="event__available-offers">
+                  <div class="event__offer-selector">
+                    <input class="event__offer-checkbox  visually-hidden" id="event-offer-luggage-1" type="checkbox" name="event-offer-luggage" checked>
+                    <label class="event__offer-label" for="event-offer-luggage-1">
+                      <span class="event__offer-title">Add luggage</span>
+                      &plus;&euro;&nbsp;
+                      <span class="event__offer-price">50</span>
+                    </label>
+                  </div>
+
+                  <div class="event__offer-selector">
+                    <input class="event__offer-checkbox  visually-hidden" id="event-offer-comfort-1" type="checkbox" name="event-offer-comfort" checked>
+                    <label class="event__offer-label" for="event-offer-comfort-1">
+                      <span class="event__offer-title">Switch to comfort</span>
+                      &plus;&euro;&nbsp;
+                      <span class="event__offer-price">80</span>
+                    </label>
+                  </div>
+
+                  <div class="event__offer-selector">
+                    <input class="event__offer-checkbox  visually-hidden" id="event-offer-meal-1" type="checkbox" name="event-offer-meal">
+                    <label class="event__offer-label" for="event-offer-meal-1">
+                      <span class="event__offer-title">Add meal</span>
+                      &plus;&euro;&nbsp;
+                      <span class="event__offer-price">15</span>
+                    </label>
+                  </div>
+
+                  <div class="event__offer-selector">
+                    <input class="event__offer-checkbox  visually-hidden" id="event-offer-seats-1" type="checkbox" name="event-offer-seats">
+                    <label class="event__offer-label" for="event-offer-seats-1">
+                      <span class="event__offer-title">Choose seats</span>
+                      &plus;&euro;&nbsp;
+                      <span class="event__offer-price">5</span>
+                    </label>
+                  </div>
+
+                  <div class="event__offer-selector">
+                    <input class="event__offer-checkbox  visually-hidden" id="event-offer-train-1" type="checkbox" name="event-offer-train">
+                    <label class="event__offer-label" for="event-offer-train-1">
+                      <span class="event__offer-title">Travel by train</span>
+                      &plus;&euro;&nbsp;
+                      <span class="event__offer-price">40</span>
+                    </label>
+                  </div>
+                </div>
+              </section>
+
+              <section class="event__section  event__section--destination">
+                <h3 class="event__section-title  event__section-title--destination">Destination</h3>
+                <p class="event__destination-description">Chamonix-Mont-Blanc (usually shortened to Chamonix) is a resort area near the junction of France, Switzerland and Italy. At the base of Mont Blanc, the highest summit in the Alps, it's renowned for its skiing.</p>
+              </section>
             </section>
           </form>`;
 }
 
 export default class NewPoint extends AbstractView {
-  #point;
-  #offers;
-  #destination;
-  constructor ({point = BLANK_POINT, offers, destination}) {
-    super();
-    this.#point = point;
-    this.#offers = offers;
-    this.#destination = destination;
-  }
-
   get template () {
-    return createNewPoint(this.#point, this.#offers, this.#destination);
+    return createNewPoint();
   }
 }
+
